// src/generated/server/worldmonitor/uap/v1/service_server.ts
export interface UapSighting {
  id: string;
  lat: number;
  lon: number;
  timestamp: number;
  source: string;
  shape: string;
  description: string;
  credibilityScore: number;
}